feat: Add onNewIntent override function for single top activity
This commit adds the `onNewIntent` override function to the activity. It is important to note that this function will only be triggered when the activity is declared as `singleTop` in the manifest. By implementing this function, the activity can handle new intents that are received while it is in the foreground, allowing it to respond appropriately to these intents.