feat: Update PhotoViewModel URI in onNewIntent
This commit adds the `viewModel.updateUncompressUri(uri)` statement to the `onNewIntent` method. The `updateUncompressUri` method of the `PhotoViewModel` is called with the URI received from another app, such as Google Chrome. This update allows the view model to store and manage the URI, facilitating further processing or display of the uncompressed image within the associated activity.