feat: Update PhotoViewModel with Work ID in onNewIntent
This commit includes the statement `viewModel.updateWorkId(request.id)` in the `onNewIntent` method. The `updateWorkId` method of the `PhotoViewModel` is called with the ID of the work request that was launched. This update enables the view model to keep track of the specific worker associated with the request, allowing it to listen for updates and handle the progress or completion of that worker.